requirements.txt
pygame==2.0.1
PyOpenGL==3.1.5
numpy==1.20.3